requirements.txt
requests
matrix-nio
schedule
pyyaml
python-magic
Pillow